+++ gpaw/cli/sbatch.py
@@ -1,18 +1,25 @@
 import subprocess
 import sys
 import os
+
 usage = """gpaw sbatch [-0] -- [sbatch options] script.py [script options]
    or: gpaw sbatch [-0] -- [sbatch options] python -m module [module options]
 """
 
 
 class CLICommand:
-    """Submit a GPAW Python script via sbatch."""
+    """Submit a GPAW Python script via sbatch.
+
+    
+   If a virtual environment is active when submitting, it will be activated for
+   the job as well.
+   """
 
     @staticmethod
     def add_arguments(parser):
         parser.usage = usage
-        parser.add_argument('-0', '--test', action='store_true')
+        parser.add_argument('-0', '--test', action='store_true',
+                            help='Print the driver script instead of submitting it.')
         parser.add_argument('arguments', nargs='*')
 
     @staticmethod
@@ -39,8 +46,8 @@ class CLICommand:
         script += ('cd $SLURM_SUBMIT_DIR\n')
         venv = os.getenv('VIRTUAL_ENV')
         if venv:
-            print("Detected virtual environment:", venv)
-            script += 'source {}/bin/activate\n'.format(venv)
+            print('Detected virtual environment:', venv)
+            script += f'source {venv}/bin/activate\n'
         script += ('OMP_NUM_THREADS=1 '
                    'mpiexec `echo $GPAW_MPI_OPTIONS` gpaw python {}\n'
                    .format(' '.join(args.arguments[i:])))